agentic_poc/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.14
langgraph==0.2.3
langchain-openai==0.1.22
chromadb==0.5.5
sentence-transformers==3.0.1
pytest==8.3.2
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
langchain-chroma==0.1.2
langchain-community==0.2.10
langchain-google-genai==1.0.8
langchain-openai==0.1.22
